+++ Valour-Bot/models/Help.cs
@@ -1,16 +1,10 @@
-using System;
 using System.ComponentModel.DataAnnotations;
 
 namespace PopeAI.Models
 {
-    public class Stat
+    public class Help
     {
         [Key]
-        public ulong Id {get; set;}
-        public ulong PlanetId {get; set;}
-        public double NewCoins {get; set;}
-        public ulong MessagesUsersSent {get; set;}
-        public ulong MessagesSent {get; set;}
-        public DateTime Time {get; set;}
+        public string Message {get; set;}
     }
 }
